# book-network-ui/src/openapi/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenApi specification - Hamidou","description":"OpenApi documentation for Spring Security","termsOfService":"Terms of service","contact":{"name":"Hamidou","url":"https://hamidou.com/course","email":"bah627320@gmail.com"},"license":{"name":"Licence name","url":"https://some-url.com"},"version":"1.0"},"servers":[{"url":"http://localhost:8088/api/v1","description":"Local ENV"},{"url":"https://aliboucoding.com/course","description":"PROD ENV"}],"security":[{"bearerAuth":[]}],"paths":{"/auth/register":{"post":{"tags":["Authentication"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationRequest"}}},"required":true},"responses":{"202":{"description":"Accepted","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/auth/authenticate":{"post":{"tags":["Authentication"],"operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/auth/activate-account":{"get":{"tags":["Authentication"],"operationId":"confirm","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"RegistrationRequest":{"type":"object","properties":{"firstname":{"type":"string","minLength":1},"lastname":{"type":"string","minLength":1},"email":{"type":"string","minLength":1},"password":{"type":"string","maxLength":2147483647,"minLength":8}},"required":["email","firstname","lastname","password"]},"AuthenticationRequest":{"type":"object","properties":{"email":{"type":"string","minLength":1},"password":{"type":"string","maxLength":2147483647,"minLength":8}},"required":["email","password"]},"AuthenticationResponse":{"type":"object","properties":{"token":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"JWT auth description","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}